refactor: purge legacy UI artifacts
