+++ app/src/main/res/values-v29/styles.xml
@@ -1,14 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <style name="Theme.HermesCeleste" parent="android:style/Theme.Material.Light.NoActionBar">
-        <item name="android:fontFamily">sans</item>
-        <item name="android:windowBackground">@color/celeste_paper</item>
-        <item name="android:statusBarColor">@color/celeste_paper</item>
-        <item name="android:navigationBarColor">@color/celeste_paper</item>
-        <item name="android:windowLightStatusBar">true</item>
-        <item name="android:windowLightNavigationBar">true</item>
-        <item name="android:colorAccent">@color/celeste_blue</item>
+    <style name="Theme.HermesCeleste" parent="@style/Theme.HermesCelesteBase">
         <item name="android:forceDarkAllowed">false</item>
-        <item name="android:windowActionModeOverlay">true</item>
     </style>
 </resources>
