+++ app/src/main/res/values/styles.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <style name="Theme.HermesCeleste" parent="android:style/Theme.Material.Light.NoActionBar">
-        <item name="android:fontFamily">sans</item>
+    <style name="Theme.HermesCelesteBase" parent="android:style/Theme.Material.Light.NoActionBar">
+        <item name="android:fontFamily">@font/inter_variable</item>
         <item name="android:windowBackground">@color/celeste_paper</item>
         <item name="android:statusBarColor">@color/celeste_paper</item>
         <item name="android:navigationBarColor">@color/celeste_paper</item>
@@ -10,4 +10,6 @@
         <item name="android:colorAccent">@color/celeste_blue</item>
         <item name="android:windowActionModeOverlay">true</item>
     </style>
+
+    <style name="Theme.HermesCeleste" parent="@style/Theme.HermesCelesteBase" />
 </resources>
